chore: extract network tab
